+++ system-parachains/coretime/coretime-polkadot/src/tests.rs
@@ -26,9 +26,9 @@ use frame_support::{
 		Get, OnInitialize,
 	},
 };
-use kusama_runtime_constants::system_parachain::coretime::TIMESLICE_PERIOD;
 use pallet_broker::{ConfigRecordOf, RCBlockNumberOf, SaleInfo};
 use parachains_runtimes_test_utils::ExtBuilder;
+use polkadot_runtime_constants::system_parachain::coretime::TIMESLICE_PERIOD;
 use sp_runtime::traits::AccountIdConversion;
 
 fn advance_to(b: BlockNumber) {